+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Fix some harmless(?) warning about material grade command arguments [#80]
+
 ## [1.4.5] - 2019-11-25
 ### Changed
 - Armor now gains Mine and Slash stats (AzureZhen) [#78, #79]